+++ packages/components/src/ProForm/type.ts
@@ -7,7 +7,6 @@ import {
   CardProps,
   CollapseProps,
   CollapsePanelProps,
-  ColProps,
 } from 'uiw';
 export interface ProFormProps {
   formDatas?: FormItemsProps[];
@@ -61,7 +60,7 @@ export interface FormItemsProps
   required?: boolean;
   // 验证规则
   rules?: RulesProps[];
-  colProps?: ColProps;
+  colstyle?: React.CSSProperties;
 }
 
 export interface OptionsProps {