+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 严格遵循 [`Semantic Versioning 2.0.0`](http://semver.org/lang/zh-CN/) 语义化版本规范。
 
-## 5.2.22(未发版)
+## 5.2.22
 
 - @uiw-admin/plugins
   - 🌟 `routes`: 路由转换使用`ast`树直接转换